move MockBalanceReporter to utils
