+++ acceptance-tests/src/test/java/tech/pegasys/web3signer/dsl/signer/Signer.java
@@ -46,6 +46,7 @@ import io.vertx.core.json.JsonObject;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.tuweni.bytes.Bytes;
+import org.web3j.protocol.Web3j;
 import org.web3j.protocol.core.Ethereum;
 import org.web3j.protocol.core.JsonRpc2_0Web3j;
 import org.web3j.protocol.http.HttpService;
@@ -72,7 +73,7 @@ public class Signer extends FilecoinJsonRpcEndpoint {
   private final Vertx vertx;
   private final String urlFormatting;
   private final Optional<ClientTlsConfig> clientTlsConfig;
-  private JsonRpc2_0Web3j signerJsonRpc;
+  private Web3j jsonRpc;
 
   public Signer(final SignerConfiguration signerConfig, final ClientTlsConfig clientTlsConfig) {
     super(JSON_RPC_PATH);
@@ -89,7 +90,7 @@ public class Signer extends FilecoinJsonRpcEndpoint {
     LOG.info("Starting Web3Signer");
     runner.start();
     final String httpUrl = getUrl();
-    signerJsonRpc = new JsonRpc2_0Web3j(new HttpService(httpUrl));
+    jsonRpc = new JsonRpc2_0Web3j(new HttpService(httpUrl));
     LOG.info("Http requests being submitted to : {} ", httpUrl);
   }
 
@@ -197,6 +198,6 @@ public class Signer extends FilecoinJsonRpcEndpoint {
   }
 
   public Ethereum jsonRpc() {
-    return signerJsonRpc;
+    return jsonRpc;
   }
 }
